requirements.txt
# LLM Providers
openai>=1.0
anthropic>=0.18
cohere>=5.0

# Core
pyyaml>=6.0
aiosqlite>=0.19
pydantic>=2.0
python-dotenv>=1.0

# Visualization & CLI
matplotlib>=3.7
click>=8.1

# Testing
pytest>=7.4
pytest-asyncio>=0.21
pytest-cov>=4.1
